backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.34.149
requests==2.32.3
httpx==0.27.2
pydicom==2.4.4
python-multipart==0.0.18
# python-magic requires libmagic:
#   Linux (EC2):   sudo yum install -y file-libs  OR  sudo apt-get install libmagic1
#   Windows (dev): pip install python-magic-bin instead of python-magic
python-magic==0.4.27
Pillow==10.4.0
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.23.8

# Load testing
locust==2.31.8
